fix(server): correct response creation and read size types
Moved createResponse method to public section of ServerEngine class to
ensure it is accessible where needed. Changed the type of bytesReadFromFd
to long int in Client.cpp to match the return type of read() and added
explicit casts to long for comparisons with nextReadSize_. This ensures
correct type handling and prevents potential issues with signed/unsigned
comparisons.

+++ srcs/Client.cpp
@@ -112,7 +112,7 @@ bool Client::hasRequestReady(void)
 	}
 
 	std::vector<char> buffer(nextReadSize_);
-	size_t bytesReadFromFd = read(pollFd_, buffer.data(), buffer.size());
+	long int bytesReadFromFd = read(pollFd_, buffer.data(), buffer.size());
 	totalBytesReadFromFd_ += bytesReadFromFd;
 
 	if (bytesReadFromFd < 0)
@@ -136,7 +136,7 @@ bool Client::hasRequestReady(void)
 	clientBuffer_.write(buffer.data(), bytesReadFromFd);
 	requestStr_ = clientBuffer_.str();
 
-	if (bytesReadFromFd != nextReadSize_)
+	if (bytesReadFromFd != (long)nextReadSize_)
 	{
 		Logger::log(Logger::DEBUG)
 			<< "hasRequestReady: bytesReadFromFd: " << bytesReadFromFd
@@ -147,7 +147,7 @@ bool Client::hasRequestReady(void)
 		return false;
 	}
 
-	if (readingPartialBody_ == true && bytesReadFromFd == nextReadSize_)
+	if (readingPartialBody_ == true && bytesReadFromFd == (long)nextReadSize_)
 	{
 		Logger::log(Logger::DEBUG)
 			<< "hasRequestReady: Body arrived incomplete and has now been "
